Update dependency versions and add new API route
Updated the version of dependency packages @lit-labs/ssr and lit to @ponomarevlad/lit-labs-ssr@3.1.5-ssr-readable-exports-npm-1 and @ponomarevlad/lit@2.8.0-ssr-readable-exports-npm-1 respectively. This change was due to the need to upgrade to the latest versions of these packages which include important bug fixes and new features.

Also, added a new API route '/api/edge/readable' in index.html and added the corresponding server-side rendering script in a newly created file - api/edge/readable.mjs. This adds a new functionality for rendering readable streams on the server side as part of an ongoing effort to enhance application performance and responsiveness.