Update lit-SSR dependencies and alter script setting
Updated lit-SSR dependencies to their latest versions. This was done to incorporate the latest features and improvements. Additionally, in line 53 of `src/page-template.mjs`, the `defer` attribute was added to the script tag to ensure scripts execute in the order they are defined and not block DOM completion. These changes improve overall performance and browser compatibility.